WTA #71: Moved ProjectDetailsDTO to :core:common module.
