+++ details/src/main/java/com/jacob/wakatimeapp/details/data/dtos/ProjectDetailsDTO.kt
@@ -1,0 +1,1 @@
+package com.jacob.wakatimeapp.details.data.dtos
