+++ search/src/main/java/com/jacob/wakatimeapp/search/data/network/SearchProjectNetworkData.kt
@@ -3,8 +3,8 @@ package com.jacob.wakatimeapp.search.data.network
 import arrow.core.Either
 import com.jacob.wakatimeapp.core.common.auth.AuthTokenProvider
 import com.jacob.wakatimeapp.core.common.data.BaseNetworkData
+import com.jacob.wakatimeapp.core.common.data.dtos.ProjectDetailsDTO
 import com.jacob.wakatimeapp.core.models.Error
-import com.jacob.wakatimeapp.search.data.network.dto.ProjectListDTO
 import com.jacob.wakatimeapp.search.data.network.mappers.toModelWithPageNumber
 import com.jacob.wakatimeapp.search.domain.models.ProjectListWithPageNumber
 import javax.inject.Inject
@@ -19,5 +19,5 @@ internal class SearchProjectNetworkData @Inject constructor(
         makeSafeApiCall(
             apiCall = { searchProjectAPI.getAllProjects(token = "Bearer $token", pageNumber) },
             methodName = ::getProjects.name,
-        ).map(ProjectListDTO::toModelWithPageNumber)
+        ).map(ProjectDetailsDTO::toModelWithPageNumber)
 }
